Fix type hint when hovering over labeled arguments of components.
This is a combination of `@react.component` PPX and compiler internals.
The PPX produces an object for the props, and the compiler generates a call to `Js_OO.unsafe_downgrade`.

The end result is that `~x` turns into `Js_OO.unsafe_downgrade(Props)#x`, which produces becomes 3 distinct values that appear to be at exactly the same source location: `Props`, `unsafe_downgrade`, and `x`.

Fixes https://github.com/rescript-lang/rescript-editor-support/issues/63