Use the grammar to highlight when to/downto are used as labels.
Labels are not handled by semantic highlighting (because there is no precise location information in the AST).
But the grammar defines to/downto as keywords, in order to handle for loops.

There are 3 cases where to/downto can be labels:
1) function definition `(~to as x) =>...`
2) function application `foo(~to=...)`
3) JSX props `<foo to=...`
(punned cases are not problematic as they are also variables, and variables are handled via semantic highlighting).

The grammar now recognises patterns of the form `to = ` and `to as` and highlights `to` as a variable.

The correctness of this relies on an observation: in `to =` and `to as`, it's not possible that `to` is the keyword of a syntactically valid for loop, as for that to be the case it must be followed by an expression.