Fix extension crash when renaming a file.
More robust solution where the file cache is set also for files with extension not `.res` or `.resi`.
In case of rename e.g. from `.res` to `.rex` an opened file notification is sent for the `.rex` file and a crash happens when trying to get the current file content.
After this change, the file content array is set also for the .rex` file, preventing the crash.

Fixes https://github.com/rescript-lang/rescript-vscode/issues/392