Emit nested document symbols.
Fixes https://github.com/rescript-lang/rescript-vscode/issues/654.

The change in https://github.com/rescript-lang/rescript-vscode/pull/629 moved to a non-deprecated representation of document symbol.
The new representation is the "Ikea" version of the old one, where children symbol are not computed for you, but need to be provided.
This PR computes and emits the tree of symbols using the `children` field.