Tests adding @rescript/core to the tests.
Just to see how things look like.
One comment is that the local file `Object.res` had to be renames as otherwise it would become unreachable.
This is a side-effect of doing `-open RescriptCore` globally, which overrides the local scope. This won't be a problem when the library is integrated in the compiler.